server/services/ai-service/requirements.txt
fastapi==0.103.1
uvicorn~=0.34.0
langchain~=0.3.20
openai==0.28.0
pydantic~=2.10.6
python-dotenv==1.0.0
requests==2.31.0
pymongo~=4.8.0